backend/requirements.txt
fastapi[standard]
openai
sqlalchemy
psycopg2-binary
pyjwt
alembic
passlib
bcrypt
anthropic
python-multipart
websockets
tavily-python